ProjetoIngles/karaoke_backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-multipart

httpx

pytubefix
pydub

numpy
rapidfuzz
